isotropic cloud moved amuse/ext -> amuse/ic; updating docstring style
